# server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBin API
  version: 1.0.0
  description: |
    REST API for OpenBin — a multi-user web app for organizing physical storage bins with QR codes.

    **Authentication**: Most endpoints require authentication via one of:
    - **httpOnly cookies** (primary) — `openbin-access` (15 min) + `openbin-refresh` (7 days), set automatically on login/register. Browsers send these with every request.
    - **API key** created via `POST /api-keys` (long-lived, per-user, format: `sk_openbin_...`) — `Authorization: Bearer <key>`. A key grants access to all locations the user belongs to.

    **Response envelopes**: All list endpoints return `{ results: T[], count: number }`.
    Error responses use `{ error: "CODE", message: "Human readable" }`.

    **Rate limiting**: All API endpoints share a global limit of 200 requests/minute.
servers:
  - url: /api
    description: API base path (proxied via nginx)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key (sk_openbin_...) via Authorization header. Cookies are the primary auth method for browsers.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - NOT_FOUND
            - FORBIDDEN
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - GONE
            - MALWARE_DETECTED
            - SCAN_UNAVAILABLE
            - SERVICE_NOT_CONFIGURED
        message:
          type: string
      example:
        error: NOT_FOUND
        message: Bin not found

    # ── Auth ───────────────────────────────────────────────
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key_prefix:
          type: string
          description: First 18 characters of the key for identification
        name:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
    ApiKeyCreateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          description: "Full API key — shown only once at creation time"
        keyPrefix:
          type: string
        name:
          type: string

    # ── Query & Execute ───────────────────────────────────
...
    PlanRestrictedError:
      type: object
      properties:
        error:
          type: string
          example: PLAN_RESTRICTED
        message:
          type: string
        upgrade_url:
          type: string
          nullable: true
          description: URL to upgrade the plan (null when not applicable)
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Not authenticated
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Name is required
...
    PlanRestricted:
      description: Feature requires Pro plan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PlanRestrictedError'
          example:
            error: PLAN_RESTRICTED
            message: This feature requires a Pro plan
            upgrade_url: https://example.com/upgrade
paths:
...
  /api-keys:
    get:
      tags: [ApiKeys]
      summary: List API keys
      description: Returns the user's active (non-revoked) API keys. Full key is never returned.
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [ApiKeys]
      summary: Create an API key
      description: |
        Creates a new API key. The full key is returned only in this response — it cannot be retrieved later.
        Maximum 10 active keys per user. Keys use the format `sk_openbin_<64 hex chars>`.
        Requires Pro plan.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                  description: A label to identify this key
      responses:
        '201':
          description: API key created (full key shown only once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PlanRestricted'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api-keys/{id}:
    delete:
      tags: [ApiKeys]
      summary: Revoke an API key
      description: Soft-revokes an API key by setting revoked_at. The key stops working immediately.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # Activity
  # ════════════════════════════════════════════════════════
...